Refactor: 뷰모델로 수정
