+++ app/src/main/res/layout/fragment_my_page.xml
@@ -47,7 +47,7 @@
 
                 <com.google.android.material.imageview.ShapeableImageView
                     android:id="@+id/iv_profile_image"
-                    glide="@{studyUser.image}"
+                    imageUrl="@{studyUser.image}"
                     android:layout_width="100dp"
                     android:layout_height="100dp"
                     app:layout_constraintEnd_toEndOf="parent"
